// manager/back/src/core/apis/agent-prod/api.ts
export interface DeployConfigModel {
    /**
     * 
     * @type {DockerField}
     * @memberof DeployConfigModel
     */
    docker: DockerField;
    /**
     * URI of the production agent
     * @type {string}
     * @memberof DeployConfigModel
     */
    uri: string;
}
/**
 * 
 * @export
 * @interface DeployJobModel
 */
export interface DeployJobModel {
    /**
     * Job id
     * @type {number}
     * @memberof DeployJobModel
     */
    id: number;
    /**
     * 
     * @type {DeployConfigModel}
     * @memberof DeployJobModel
     */
    config?: DeployConfigModel;
}
/**
 * 
 * @export
 * @interface DockerComposeField
 */
export interface DockerComposeField {
    /**
     * Path where the docker-compose.yml file is
     * @type {string}
     * @memberof DockerComposeField
     */
    path?: string;
}
/**
 * 
 * @export
 * @interface DockerField
 */
export interface DockerField {
    /**
     * 
     * @type {DockerComposeField}
     * @memberof DockerField
     */
    compose?: DockerComposeField;
}
// ...
export class ProductionAgentApi extends BaseAPI {
    /**
     * Deploy a project following a configuration
     * @param {DeployJobModel} deployJobModel 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof ProductionAgentApi
     */
    public productionAgentBuild(deployJobModel: DeployJobModel, options?: any) {
        return ProductionAgentApiFp(this.configuration).productionAgentBuild(deployJobModel, options).then((request) => request(this.axios, this.basePath));
    }
}